liquid_server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==2.9.1
liquidtrading-python==0.1.0
python-dotenv==1.0.1
cryptography==43.0.1
gunicorn==21.2.0
